optimize cluster runtime
update reviews: upload container image to registry by driver when install app
